Source ‘bash_init.sh’ for ‘require_cmd’.
The rule is that only commands without shebang has implicit sourcing of this initialization file. However, the way it is implemented in TextMate 1.x is by setting BASH_ENV in the environment for the command being executed, so if this command later runs a bash script, the variable is still in effect.

2.0 changes this to the strict interpretation of the rule so we need to explicitly source the initialization file if we use any of the functions it declares.

This commit still provides compatibility with TextMate 1.x. Ideally we add the command requirements to the bundle items (requires 2.0).