Add “Open Package” (⇧⌘D), uses current line’s import path
If the current line is not an import statement then it will prompt the user for the package to open. Default dialog text is the contents of the search clipboard.

The package is located via ‘GOPATH’ and ‘GOROOT’ (obtained via ‘go env’).

If the package contains a single go file (excluding tests) then that file will be opened, otherwise the package is opened as a new project.