Keys in map literals no longer matched as having zero indent
The rule was intended to match labels, which look the same as key literals.

Now ‘foo:’ is only matched as a label if it has no leading whitespace.

Given the simplicity of the Go grammar, we could actually match map literals and give them a different scope, and thus have different indent patterns for when we are inside a map literal.